Memory crash resolution: Identified likely circular reference in supply of muscatConversion to marcConversion
Analysis using xdebug_debug_zval() indicates a circular reference in loadLookupTable by use of $this->muscatConversion.
Currently, marcConversion receives the parent object $this->muscatConversion.
marcConversion->loadLookupTable() is run once but presumably cannot free this resource.
The design is currently poor, as $this->muscatConversion is only being passed in so that its methods are available to both marcConversion and reports.
However, the common methods really relate to MARC conversion -related aspects, such as diacritics, asterisk escaping, ISBN.
Thus the class hierarchy should be changed so that reports is passed the knowledge of marcConversion.
This first commit passes this in, to enable the methods to be moved.