Add replacement transliteration table loading algorithm, fixing many edge-case issues and missing implementation
Previously, the algorithm added in all records with lang=Russian, and then pared this down by deleting non-relevant shards (e.g. if *in contains a *tg/*lang=Russian then the whole *in section is deleted).

There were two problems with this approach:
1) Fundamentally that it doesn't cope with things like *nt where a whole record is in English but then there's *nt=BGNRus - because the shard will never be there at all from the start.
2) The rules overlap. A rule might delete a *in with *lang=English block, but if that contains a parallel title, we have lost that block.

This new algorithm copies in *every* shard, sets the language of each shard to default to the record language, and then progressively works through the various rules with more specific language where necessary. At the end, there is then a deletion of everything that's not relevant.