{1|7}xx: Add support for *nt=None
This is achieved by use of a new function, transliterationDisabledNt().
This takes the language mode, which is therefore passed in to all instances of main().
If a parallel *nt (i.e. $path + '/nt') exists, and equals 'None', then main returns false.
Each caller of the types of author groups, may thus end up with an empty (false) line registered.
However, further modifications of the line must be avoided, e.g. adding "(alternative name)", hence the use of if ($line) {} in several cases.
Simply skipping the line would end up with mismatches, hence the creation of an empty line which the 880 then discards.
The 880 handler then detects an empty line, and skips this line.
However, the 880 counter does not increment so that the counter is also kept in sync.