Convert getTransliterableFullStringsInBrackets to late binding, fixing bug in 5884120e356be6d033bf5073b553682e50805b39, fixing failing test #1022
transliteration::getTransliterableFullStringsInBrackets () has a dependency on catalogue_processed, but $this->transliteration is instantiated before the start of input so catalogue_processed does not exist.
This commit sets getTransliterableFullStringsInBrackets () to be late-bound, so that the first time it is called, which is during conversion of a record, it loads the data, and then caches it.