Redefine hostRecordId as properly meaning either *kg or implicit title match
Previously $this->hostRecord was derived only from *kg.
However there were still cases of *art/*j that still use the implicit title match (*t), which were not being handled.
The result of this was 773 was not being created in many cases, and a hard-coded 852 also resulted.

Also determined that *art/*in is believed always to have a *kg when there is a parent.