852 If there is a host record, and the current location has been looked up from location=Periodical, skip item record created for this location
I.e. there is a host record from which this location has been derived.
Essentially, this scenario is where a host (e.g. *kg) and location=Periodical are considered to be the same location, rather than an additional location.
There may be other locations present.

The core change here is in itemRecordsCreation, where the host record check is moved up to apply to all record types, rather than just a more basic check in *art