Fix failing test for transliteration handling of [strings], believed arising from some phasing bug
The record (and other records needing the same feature) were failing during import.
But these were not failing when re-running tests after import.
Believed therefore some kind of phasing problem.
A previous phasing bug aimed at dealing with this seems not to have solved it.

Therefore, have reimplemented this using a static file load, which should not be phasing-dependent.